+++ src/main/java/com/openclassrooms/rental/repository/RentalRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RentalRepository extends CrudRepository<RentalEntity, Integer> {
+public interface RentalRepository extends CrudRepository<RentalEntity, Long> {
 }
