+++ src/main/java/com/openclassrooms/rental/service/query/UserQueryService.java
@@ -22,7 +22,7 @@ public class UserQueryService implements UserDetailsService {
         this.userMapper = userMapper;
     }
 
-    public UserResponse getUserById(Integer id) throws NoSuchElementException {
+    public UserResponse getUserById(Long id) throws NoSuchElementException {
         UserEntity user = userRepository.findById(id)
                 .orElseThrow(() -> new NoSuchElementException("User not found"));
         return userMapper.toDto(user);
